adapter rename to spark-cde
