+++ dbt/adapters/spark_cde/__version__.py
@@ -1,0 +1,1 @@
+version = "1.1.0"
