+++ dbt/adapters/spark_cde/column.py
