+++ dbt/adapters/spark_cde/livysession.py
