+++ dbt/adapters/spark_cde/session.py
