--- dbt/adapters/spark_livy/__init__.py
@@ -1,11 +1,0 @@
-from dbt.adapters.spark_livy.connections import SparkConnectionManager  # noqa
-from dbt.adapters.spark_livy.connections import SparkCredentials
-from dbt.adapters.spark_livy.relation import SparkRelation  # noqa
-from dbt.adapters.spark_livy.column import SparkColumn  # noqa
-from dbt.adapters.spark_livy.impl import SparkAdapter
-
-from dbt.adapters.base import AdapterPlugin
-from dbt.include import spark_livy
-
-Plugin = AdapterPlugin(
-    adapter=SparkAdapter, credentials=SparkCredentials, include_path=spark_livy.PACKAGE_PATH)
