--- dbt/adapters/spark_livy/__version__.py
@@ -1,1 +1,0 @@
-version = "1.1.2"
