+++ dbt/include/spark_cde/macros/materializations/incremental/incremental.sql
@@ -1,4 +1,4 @@
-{% materialization incremental, adapter='spark_livy' -%}
+{% materialization incremental, adapter='spark_cde' -%}
 
   {#-- Validate early so we don't run SQL if the file_format + strategy combo is invalid --#}
   {%- set raw_file_format = config.get('file_format', default='parquet') -%}
