+++ dbt/include/spark_cde/macros/materializations/snapshot.sql
@@ -1,4 +1,4 @@
-{% macro spark_livy__snapshot_hash_arguments(args) -%}
+{% macro spark_cde__snapshot_hash_arguments(args) -%}
     md5({%- for arg in args -%}
         coalesce(cast({{ arg }} as string ), '')
         {% if not loop.last %} || '|' || {% endif %}
@@ -6,13 +6,13 @@
 {%- endmacro %}
 
 
-{% macro spark_livy__snapshot_string_as_time(timestamp) -%}
+{% macro spark_cde__snapshot_string_as_time(timestamp) -%}
     {%- set result = "to_timestamp('" ~ timestamp ~ "')" -%}
     {{ return(result) }}
 {%- endmacro %}
 
 
-{% macro spark_livy__snapshot_merge_sql(target, source, insert_cols) -%}
+{% macro spark_cde__snapshot_merge_sql(target, source, insert_cols) -%}
 
     merge into {{ target }} as DBT_INTERNAL_DEST
     using {{ source }} as DBT_INTERNAL_SOURCE
@@ -49,12 +49,12 @@
 {% endmacro %}
 
 
-{% macro spark_livy__post_snapshot(staging_relation) %}
+{% macro spark_cde__post_snapshot(staging_relation) %}
     {% do adapter.drop_relation(staging_relation) %}
 {% endmacro %}
 
 
-{% macro spark_livy__create_columns(relation, columns) %}
+{% macro spark_cde__create_columns(relation, columns) %}
     {% if columns|length > 0 %}
     {% call statement() %}
       alter table {{ relation }} add columns (
