+++ dbt/include/spark_cde/macros/materializations/view.sql
@@ -1,3 +1,3 @@
-{% materialization view, adapter='spark_livy' -%}
+{% materialization view, adapter='spark_cde' -%}
     {{ return(create_or_replace_view()) }}
 {%- endmaterialization %}
