+++ dbt/include/spark_cde/macros/utils/any_value.sql
@@ -1,4 +1,4 @@
-{% macro spark_livy__any_value(expression) -%}
+{% macro spark_cde__any_value(expression) -%}
     {#-- return any value (non-deterministic)  --#}
     first({{ expression }})
