+++ dbt/include/spark_cde/macros/utils/assert_not_null.sql
@@ -2,7 +2,7 @@
   {{ return(adapter.dispatch('assert_not_null', 'dbt')(function, arg)) }}
 {%- endmacro %}
 
-{% macro spark_livy__assert_not_null(function, arg) %}
+{% macro spark_cde__assert_not_null(function, arg) %}
 
     coalesce({{function}}({{arg}}), nvl2({{function}}({{arg}}), assert_true({{function}}({{arg}}) is not null), null))
