+++ dbt/include/spark_cde/macros/utils/bool_or.sql
@@ -4,7 +4,7 @@
   -- https://spark.apache.org/docs/latest/api/sql/index.html#max
 #}
 
-{% macro spark_livy__bool_or(expression) -%}
+{% macro spark_cde__bool_or(expression) -%}
 
     max({{ expression }})
