+++ dbt/include/spark_cde/macros/utils/split_part.sql
@@ -1,4 +1,4 @@
-{% macro spark_livy__split_part(string_text, delimiter_text, part_number) %}
+{% macro spark_cde__split_part(string_text, delimiter_text, part_number) %}
 
     {% set delimiter_expr %}
